move version methods to package; fix lint issues
